Split up `load_settings` and only test the individual pieces
