+++ cnd/src/lib.rs
@@ -14,7 +14,6 @@ pub mod comit_client;
 pub mod comit_i_routes;
 pub mod http_api;
 pub mod libp2p_bam;
-pub mod load_settings;
 pub mod logging;
 pub mod network;
 pub mod seed;
